pyproject.toml
[tool.poetry]
name = "bilimai"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.4"
djangorestframework = "^3.14.0"
djoser = "^2.2.0"
langchain = "^0.0.270"
openai = "^0.27.8"
chromadb = "^0.4.6"
jq = "^1.4.1"
pinecone-client = "^2.2.2"
tiktoken = "^0.4.0"
nest-asyncio = "^1.5.7"
bs4 = "^0.0.1"
google-search-results = "^2.4.2"
googletrans = "^3.0.0"
translators = "^5.8.2"
googlesearch-python = "^1.2.3"
google = "^3.0.0"
certifi = "^2023.7.22"
django-filter = "^23.2"
pillow = "^10.0.0"
django-cors-headers = "^4.2.0"
django-phonenumber-field = {extras = ["phonenumberslite"], version = "^7.1.0"}
unstructured = "^0.10.11"
gunicorn = "^21.2.0"
whitenoise = "^6.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
